app/requirements.txt
uvicorn
fastapi
openai
python-dotenv
azure-ai-projects
azure-identity
azure-search-documents
sqlalchemy
mysql-connector-python
gunicorn